requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
astroid==1.5.3
attrs==17.2.0
Automat==0.6.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.0.3
cssselect==1.0.1
defusedxml==0.5.0
Django==1.11.6
django-allauth==0.34.0
django-filter==1.1.0
django-rest-auth==0.9.2
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
hyperlink==17.1.1
hypothesis==3.23.0
idna==2.6
incremental==17.5.0
isort==4.2.15
itypes==1.1.0
Jinja2==2.9.6
lazy-object-proxy==1.3.1
lxml==4.1.0
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.6
parsel==1.2.0
psycopg2==2.7.3.1
pyasn1==0.3.7
pyasn1-modules==0.0.8
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.7.4
Pympler==0.5
pyOpenSSL==17.2.0
PySocks==1.6.7
python3-openid==3.1.0
pytz==2017.2
queuelib==1.4.2
requests==2.18.4
requests-oauthlib==0.8.0
Scrapy==1.4.0
selenium==3.6.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
uritemplate==3.0.0
urllib3==1.22
w3lib==1.18.0
wrapt==1.10.11
zope.interface==4.4.3
